Promote Timer to an internal util class
